\name{ex0726}
\alias{ex0726}
\docType{data}
\title{Decline in Male Births}
\description{
  The data are on the proportion of male birts in Denmark, The
  Netherlands, Canada and the United States for a number of yeras.
  Notice that the proportions for Canada and the United States are only
  provided for the years 1970 to 1990, while Denmark and The Netherlands
  have data listed for 1950 to 1994.
}
\usage{ex0726}
\format{
  A data frame with 45 observations on the following 5 variables.
  \describe{
    \item{\code{Year}}{year of observation}
    \item{\code{Denmark}}{male birth rate of Denmark for given year}
    \item{\code{Netherlands}}{male birth rate of The Netherlands for given year}
    \item{\code{Canada}}{male birth rate of Canada for given year}
    \item{\code{Usa}}{male birth rate of the United States for given year}
  }
}
\source{
  Ramsey, F.L. and Schafer, D.W. (2002). \emph{The Statistical Sleuth: A
    Course in Methods of Data Analysis (2nd ed)}, Duxbury.
}
\references{
  
  Davis, D.L., Gottlieb, M.B. and Stampnitzky, J.R. (1998).  Reduced
  ratio of male to female births in several industrial countries,
  \emph{Journal of the American Medical Association} \bold{279}(13):
  1018--1023.
  
}
\examples{
str(ex0726)
}
\keyword{datasets}
